Refactor RelationUtil. update mzLib
