+++ ProteoformSuiteGUI/packages.config
@@ -5,7 +5,7 @@
   <package id="EntityFramework" version="6.1.3" targetFramework="net452" />
   <package id="GraphSharp" version="1.1.0.0" targetFramework="net452" />
   <package id="MathNet.Numerics" version="3.17.0" targetFramework="net452" />
-  <package id="mzLib" version="1.0.153" targetFramework="net452" />
+  <package id="mzLib" version="1.0.154" targetFramework="net452" />
   <package id="QuickGraph" version="3.6.61119.7" targetFramework="net452" />
   <package id="WPFExtensions" version="1.0.0" targetFramework="net452" />
   <package id="Zlib.Portable" version="1.11.0" targetFramework="net452" />
