+++ ProteoformSuiteInternal/ProteoformCommunity.cs
@@ -322,7 +322,7 @@ namespace ProteoformSuiteInternal
 
             foreach (Proteoform p in Lollipop.proteoform_community.decoy_proteoforms.Values.SelectMany(d => d))
             {
-                p.relationships.Clear();
+                p.relationships.RemoveAll(r => r.relation_type == ProteoformComparison.ExperimentalTheoretical || r.relation_type == ProteoformComparison.ExperimentalDecoy)
             }
 
             relations_in_peaks.RemoveAll(r => r.relation_type == ProteoformComparison.ExperimentalTheoretical || r.relation_type == ProteoformComparison.ExperimentalDecoy);
