+++ ProteoformSuiteInternal/packages.config
@@ -5,6 +5,6 @@
   <package id="DocumentFormat.OpenXml" version="2.5" targetFramework="net452" />
   <package id="MathNet.Numerics" version="3.17.0" targetFramework="net452" />
   <package id="Microsoft.Office.Interop.Excel" version="15.0.4795.1000" targetFramework="net452" />
-  <package id="mzLib" version="1.0.153" targetFramework="net452" />
+  <package id="mzLib" version="1.0.154" targetFramework="net452" />
   <package id="Zlib.Portable" version="1.11.0" targetFramework="net452" />
 </packages>
