+++ Test/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="DocumentFormat.OpenXml" version="2.5" targetFramework="net452" />
   <package id="MathNet.Numerics" version="3.17.0" targetFramework="net452" />
-  <package id="mzLib" version="1.0.153" targetFramework="net452" />
+  <package id="mzLib" version="1.0.154" targetFramework="net452" />
   <package id="NUnit" version="3.6.1" targetFramework="net452" />
   <package id="NUnit3TestAdapter" version="3.7.0" targetFramework="net452" />
   <package id="Zlib.Portable" version="1.11.0" targetFramework="net452" />
